Move custom metrics to common directory (src/app/lib/utilities) so that it can reused elsewhere
