+++ src/server/index.test.jsx
@@ -9,8 +9,8 @@ import routes from '../app/routes';
 import { localBaseUrl } from '../testHelpers/config';
 import services from './utilities/serviceConfigs';
 import * as renderDocument from './Document';
-import sendCustomMetrics from './utilities/customMetrics';
-import { NON_200_RESPONSE } from './utilities/customMetrics/metrics.const';
+import sendCustomMetrics from '#lib/utilities/customMetrics';
+import { NON_200_RESPONSE } from '#lib/utilities/customMetrics/metrics.const';
 import loggerMock from '#testHelpers/loggerMock';
 import { ROUTING_INFORMATION } from '#lib/logger.const';
